fixed double-call issues with event-cbs
when an error occurs in an event-cb, before it would send the error up
the stack. if :catch-app-errors t was set, then the error would be
caught again and sent back into the event-cb. this was causing the
event-cb to be fired twice. not good.

i introduced a change in catch-app-errors and added a new macro
`run-event-cb` which tracks if an event has been thrown by the event-cb,
and if so, doesn't re-fire the event-cb with that event again.

very nice addition. implemented run-event-cb in every callback that
fires events manually, works well.

also, added a test in dns to make sure event-cb only gets fired once.