Avoid event loop cruelty upon unhandled conditions in callbacks.
This is for the case when catch-app-errors is set to false (the default).

Provide restarts for aborting the callback and exiting the current event
loop safely (without unwinding the stack through uv_run() call).

Use *sldb-quit-restart* to set the default debugger 'quit' action when
SWANK is available.  When using REPL integration, default to aborting
the callback; when using plain event loop, default to exiting it
safely.