changing the error handling system
this can be considered a breaking change, per issue #108 as well as
discussions at https://github.com/orthecreedence/blackbird/issues/8. the
idea being that cl-async no longer catches all errors and routes them to
the event-cb callbacks, but instead now looks for event-info conditions
(and all derivatives) and sends *only those* to the event-cb, leaving
all other errors to bubble to the top (debugger).

note that errors only bubble to the top if :catch-app-errors is nil! if
t, *all* errors are still caught, with one important change:

if :caught-errors is given (a function of one argument) when starting
the event loop, then all caught errors are sent to this function to do
what you will with them. in other words, event-cb is no longer the
dumping ground for all conditions when catching errors, now there is a
separate place. note that :caught-errors is only used when
:catch-app-errors is T.