Acknowledge auto-init for libssl1 (#154)
The function SSL_library_init is not available anymore in
libssl 1.0 and onwards.  The library uses another mechanism to
initialise itself.

It does not seem that there is an easy way to determine the version of
libssl we have loaded, we therefore rely on the presence of a
SSL_library_init foreign symbol to determine wether that function must
be called or not.

While libssl documentation claims the function SSL_load_error_strings
should be present in libssl 1.0 and onwards, it is not, at least on
Debian Stretch.  We use the same strategy as for SSL_library_init to
avoid a spurious call.

Reference: https://www.openssl.org/docs/man1.1.0/ssl/OPENSSL_init_ssl.html